build: refactor for configuration cache compatibility
